Rename AvbrytAktivitet til TypeScript
